added confirmation dialog for course removal

+++ src/app/courses/containers/courses/courses.component.ts
@@ -1,12 +1,13 @@
 import { AsyncPipe, NgIf } from '@angular/common';
 import { Component } from '@angular/core';
 import { MatCardModule } from '@angular/material/card';
-import { MatDialog, MatDialogModule } from '@angular/material/dialog';
+import { MatDialog } from '@angular/material/dialog';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { MatSnackBar, MatSnackBarModule } from '@angular/material/snack-bar';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Observable, catchError, of } from 'rxjs';
+import { ConfirmationDialogComponent } from '../../../shared/components/confirmation-dialog/confirmation-dialog.component';
 import { ErrorDialogComponent } from '../../../shared/components/error-dialog/error-dialog.component';
 import { CoursesListComponent } from '../../components/courses-list/courses-list.component';
 import { Course } from '../../models/course';
@@ -19,11 +20,11 @@ import { CoursesService } from '../../services/courses.service';
     AsyncPipe,
     NgIf,
     MatCardModule,
-    MatDialogModule,
     MatProgressSpinnerModule,
     MatSnackBarModule,
     MatToolbarModule,
     CoursesListComponent,
+    ConfirmationDialogComponent,
   ],
   templateUrl: './courses.component.html',
   styleUrl: './courses.component.scss',
@@ -65,17 +66,27 @@ export class CoursesComponent {
   }
 
   onRemove(course: Course) {
-    this.coursesService.remove(course._id).subscribe({
-      next: () => {
-        this.refresh();
-        this._snackBar.open('Course removed successfully', 'X', {
-          duration: 5000,
-          verticalPosition: 'top',
-          horizontalPosition: 'center',
-        });
-      },
-      error: () => {
-        this.onError('Error removing course.');
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Are you sure you want to remove this course?',
+    });
+
+    dialogRef.afterClosed().subscribe({
+      next: (result) => {
+        if (result) {
+          this.coursesService.remove(course._id).subscribe({
+            next: () => {
+              this.refresh();
+              this._snackBar.open('Course removed successfully', 'X', {
+                duration: 5000,
+                verticalPosition: 'top',
+                horizontalPosition: 'center',
+              });
+            },
+            error: () => {
+              this.onError('Error removing course.');
+            },
+          });
+        }
       },
     });
   }
